Rename source files to lowercase names Windows may be case insensitive, but real operating systems aren't.
