+++ src/components/ppu.c
@@ -1,4 +1,4 @@
-#include "PPU.h"
+#include "ppu.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
